requirements.txt
asgiref==3.8.1
click==8.1.7
cv==1.0.0
Django==5.0.3
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
image==1.5.33
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.1
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
